Adds basic verse selection capability
The user can click on the text of a verse to highlight it and add it to
the list of referenced verses. These are displayed above the text input
boxes and also stored in the sqlite database.

The Javascript was moved into a new file and referenced from the
index.html page.

I have run out of free quote from Cursor tonight. When I next can work
with it I need to fix a bug where clicking between the lines of verses
causes the first verse in a passage to be selected rather than the verse
where the lines were clicked.

Beyond that I need to begin refactoring the server and javascript code
into more files and functions.